+++ src/main/java/com/pintoss/auth/api/payment/PaymentController.java
@@ -1,4 +1,4 @@
-package com.pintoss.auth.core.payment.api;
+package com.pintoss.auth.api.payment;
 
 import com.pintoss.auth.common.dto.ApiResponse;
 import com.pintoss.auth.core.payment.application.PaymentService;
